More install scripts & README adjust

+++ bin/install/neovim.sh
@@ -4,12 +4,19 @@ paru -S --needed neovim-nightly-bin
 sudo pacman -S --needed xclip python
 
 # Install plug.vim
-sh -c 'curl -fLo "${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim'
+if [ ! -f "${XDG_DATA_HOME:-$HOME/.local/share}/nvim/site/autoload/plug.vim" ]; then
+        sh -c 'curl -fLo "${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim'
+fi
 
 # Make the python virtual environment
-mkdir -p $XDG_CONFIG_HOME/pyvirtualenvs
-python3 -m venv $XDG_CONFIG_HOME/pyvirtualenvs/neovim
-source $XDG_CONFIG_HOME/pyvirtualenvs/neovim/bin/activate
-pip install pynvim
-pip install black
-deactivate
+if [ ! -f "$XDG_CONFIG_HOME/pyvirtualenvs/neovim" ]; then
+        mkdir -p $XDG_CONFIG_HOME/pyvirtualenvs
+        python3 -m venv $XDG_CONFIG_HOME/pyvirtualenvs/neovim
+        source $XDG_CONFIG_HOME/pyvirtualenvs/neovim/bin/activate
+        pip install pynvim
+        pip install black
+        deactivate
+fi
+
+# Install language servers
+sudo pacman -S --needed rust-analyzer pyright gopls lua-language-server typescript-language-server texlab
